+++ lib/td_dd_web/swagger_definitions.ex
@@ -17,7 +17,6 @@ defmodule TdDdWeb.SwaggerDefinitions do
           description [:string, :null], "Data Structure description"
           type [:string, :null], "Data Structure type (csv, table...)"
           ou [:string, :null], "Data Structure organizational unit"
-          lopd [:string, :null], "Data Structure lopd level"
           last_change_at :string, "Data Structure last updated at"
           inserted_at :string, "Data Structure creation date"
           data_fields Schema.ref(:DataFields)
@@ -31,12 +30,10 @@ defmodule TdDdWeb.SwaggerDefinitions do
         description: "Data Structure description",
         type: "Csv",
         ou: "General Management",
-        lopd: "1",
         inserted_at: "2018-05-08T17:17:59.691460",
         data_fields: [],
         metadata: %{"description" => "last description",
                     "ou" => "Super Management",
-                    "lopd" => "EM",
                     "last_change_at" => "2018-05-08T17:17:59.691460"}
         }
       end,
@@ -50,7 +47,6 @@ defmodule TdDdWeb.SwaggerDefinitions do
               description :string, "Data Structure description"
               type :string, "Data Structure type (csv, table...)"
               ou :string, "Data Structure organizational unit"
-              lopd :string, "Data Structure lopd level"
             end
           end)
         end
@@ -61,7 +57,6 @@ defmodule TdDdWeb.SwaggerDefinitions do
             properties do
               description :string, "Data Structure description"
               ou :string, "Data Structure organizational unit"
-              lopd :string, "Data Structure lopd level"
             end
           end)
         end