+++ priv/repo/migrations/20181119112023_add_df_content_to_data_structure.exs
@@ -1,0 +1,11 @@
+defmodule TdDd.Repo.Migrations.AddDfContentToDataStructure do
+  use Ecto.Migration
+
+  def change do
+    alter table(:data_structures) do
+      add :df_name, :string
+      add :df_content, :map
+      remove :lopd
+    end
+  end
+end