+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [2.8.1] 2018-11-19
+
+### Added
+
+- [TD-1062] Support for Dynamic Forms in data structures
+
+### Removed
+
+- Remove LOPD field from data structures
+
 ## [2.8.0] 2018-11-15
 
 ### Added